+++ packages/widgetbook/lib/src/navigation/nodes/nodes.dart
@@ -1,5 +1,7 @@
 export 'widgetbook_category.dart';
 export 'widgetbook_component.dart';
 export 'widgetbook_folder.dart';
+export 'widgetbook_node.dart';
 export 'widgetbook_package.dart';
+export 'widgetbook_root.dart';
 export 'widgetbook_use_case.dart';