+++ packages/widgetbook/lib/src/navigation/widgets/navigation_tree_node.dart
@@ -1,20 +1,19 @@
 import 'package:flutter/material.dart';
 
-import '../navigation.dart';
+import '../nodes/nodes.dart';
+import 'navigation_tree_tile.dart';
 
 class NavigationTreeNode extends StatefulWidget {
   const NavigationTreeNode({
     super.key,
-    required this.data,
-    this.level = 0,
+    required this.node,
     this.selectedNode,
     this.onNodeSelected,
   });
 
-  final int level;
-  final NavigationTreeNodeData data;
-  final NavigationTreeNodeData? selectedNode;
-  final ValueChanged<NavigationTreeNodeData>? onNodeSelected;
+  final WidgetbookNode node;
+  final WidgetbookNode? selectedNode;
+  final ValueChanged<WidgetbookNode>? onNodeSelected;
 
   @override
   State<NavigationTreeNode> createState() => _NavigationTreeNodeState();
@@ -25,52 +24,52 @@ class _NavigationTreeNodeState extends State<NavigationTreeNode> {
 
   @override
   void initState() {
-    isExpanded = widget.data.isInitiallyExpanded;
     super.initState();
+
+    isExpanded = widget.node.isInitiallyExpanded;
   }
 
   @override
   Widget build(BuildContext context) {
+    const animationDuration = Duration(
+      milliseconds: 200,
+    );
+
     return Column(
       children: [
-        NavigationTreeItem(
-          data: widget.data,
+        NavigationTreeTile(
+          node: widget.node,
           isExpanded: isExpanded,
-          level: widget.level,
-          isSelected: widget.data.path == widget.selectedNode?.path,
+          isSelected: widget.node.path == widget.selectedNode?.path,
           onTap: () {
-            setState(() {
-              isExpanded = !isExpanded;
-            });
-            if (widget.data.isSelectable) {
-              widget.onNodeSelected?.call(widget.data);
-            }
+            setState(() => isExpanded = !isExpanded);
+            widget.onNodeSelected?.call(widget.node);
           },
         ),
-        ClipRect(
-          child: AnimatedSlide(
-            duration: const Duration(milliseconds: 200),
-            curve: Curves.easeInOut,
-            offset: Offset(0, isExpanded ? 0 : -1),
-            child: AnimatedAlign(
-              duration: const Duration(milliseconds: 200),
+        if (widget.node.children != null)
+          ClipRect(
+            child: AnimatedSlide(
+              duration: animationDuration,
               curve: Curves.easeInOut,
-              alignment: Alignment.bottomCenter,
-              heightFactor: isExpanded ? 1 : 0,
-              child: ListView.builder(
-                physics: const NeverScrollableScrollPhysics(),
-                itemCount: widget.data.children.length,
-                shrinkWrap: true,
-                itemBuilder: (context, index) => NavigationTreeNode(
-                  data: widget.data.children[index],
-                  level: widget.level + 1,
-                  selectedNode: widget.selectedNode,
-                  onNodeSelected: widget.onNodeSelected,
+              offset: Offset(0, isExpanded ? 0 : -1),
+              child: AnimatedAlign(
+                duration: animationDuration,
+                curve: Curves.easeInOut,
+                alignment: Alignment.topCenter,
+                heightFactor: isExpanded ? 1 : 0,
+                child: ListView.builder(
+                  physics: const NeverScrollableScrollPhysics(),
+                  itemCount: widget.node.children!.length,
+                  shrinkWrap: true,
+                  itemBuilder: (context, index) => NavigationTreeNode(
+                    node: widget.node.children![index],
+                    selectedNode: widget.selectedNode,
+                    onNodeSelected: widget.onNodeSelected,
+                  ),
                 ),
               ),
             ),
           ),
-        ),
       ],
     );
   }